- updated TraceViewerConnection project - created a QtCreator project for Windows x86 and x64

+++ TraceClientCore/sources/TraceNxStreamReader.cpp
@@ -130,8 +130,8 @@ namespace TraceClientCore
                 pTraceData->TickCount() = m_ReadBuffer.GetBufferAs<char>();
             }
 
-            NYXTRACE(0x0, L"Time Ref in seconds : " << TimeRefInSeconds << L" - tick count ref : " << TickCountRef );
-            NYXTRACE(0x0, L"Module time ref in seconds : " << ModuleTimeRefInSeconds << L" - module tick count ref : " << ModuleTickCountRef );
+            NYXTRACE(0x0, L"Time Ref in seconds : " << (int)TimeRefInSeconds << L" - tick count ref : " << (int)TickCountRef );
+            NYXTRACE(0x0, L"Module time ref in seconds : " << (int)ModuleTimeRefInSeconds << L" - module tick count ref : " << (int)ModuleTickCountRef );
             NYXTRACE(0x0, L"Trace tick count : " << pTraceData->TickCount().c_str() );
             
             // data
